chore(window-title): window title to be dynamic
- needs to follow the page that it opens, and not all just Coursemology
- follows the last crumb
