+++ client/app/lib/hooks/router/dynamicNest/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_WINDOW_TITLE = 'Coursemology';
